Bug 6870 [enh] - allow opac-userupdate.pl to use the patron's email as 'from' address
Patch makes it so that Koha sends patron update emails from the first valid patron email address, instead of from the KohaAdmin or branch email address.

This patch should make it so that libraries can directly reply to patrons regarding patron updates.

To test:

Create a patron with no email address, do a patron update and observe the to: and from: - they should both be the KohaAdminEmail address.

Add an email address to that same patron, and send a patron update. Observe the to: and from: - the from: should be the first valid email address of the patron, to should be the KohaAdminEmail address.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Ian Walls <ian.walls@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>