Bug 6901: Add classes to XSLT detail view in STAFF
This patch adds css classes to the fields in XSLT intranet views.
Use OpacUserCss and IntranetUserCss system preferences for altering the styling of
and hiding fields.

The css classes are named like the labels with additional underscores
where necessary.

Example:
.publisher {
display:none;
}

.other_title {
font-weight: bold;
}

Note: This patch also fixes the display of additional authors in field 880 (original
scripts like Hebrew). This small additional problem was found as a side effect of testing, where the
existence of an author in 880 would break the XSLT.

Also moves some CSS from a style attribute into the CSS file for 880 fields.

Signed-off-by: Liz Rea <lrea@nekls.org>

-- minor commit message edits (requested by KF)