Bug 6323 -- Error handling cleanup for moveitem.tt
Added button to retry attaching item, added default bib view handling, cosmetic changes to the error display (buttons now in error div).

Incorporates changes that should satisfy QA.

To test:
Find a bib with multiple items and grab one of the barcodes.
Go to a second bib and click Edit -> Attach item
Paste your barcode in there and click "Submit"
Click the "Try again with a different barcode" button. It should take you back to the barcode entry screen.

Also try with a non-existent barcode and verify that the button appears and works as intended.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>