Bug 6553: parentheses in authors break tracings
If an author's name includes parentheses or other escape-characters, the tracing links
produced in the OPAC and staff client fail in XSLT mode.  This patch wraps the author names
in double-quotes, providing the necessary escapement.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Test plan:
- Create a new framework, delete Thesaurus from 700 and other fields
- Create a new record in this framework
- Add 700$a - use parentheses
- Check the search links in OPAC and staff detail pages
> before patch link was broken, after it works

Note: Links are only broken when no authorities are being used.
When using authorities the link uses the auth number instead of the text
for linking, so the parentheses don't matter for search.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>