Bug 5680: Follow Up for Order cancelling improvement : the same for parcel.pl
The same scope than in basket.pl

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Detailed information about possible deletion of orders on parcel page.

1) Delete order for record with 1 other item and item level hold
OK. Record, item and hold remain, only on order item is deleted.

2) Delete order for ordered record with 1 item and title level hold
OK. Deletion is not possible, before hold is cancelled.

3) Delete order for new record with only items connected to this order
OK. Cancelling the order deletes record and item or only order,
depending on link used.

4) Delete order for record with subscription.
OK. Subscription remains untouched, item is deleted.

Also checked:
AcqCreateItem = on receive / in cataloging
Links show and it's possible to delete the order or only the record
using different links.

Comments:
There is a table sorter option for the last column with the delete links.
I am not sure it is useful here, but it does work.

Before patch is applied, the links use 'cancel' instead of 'delete'.
I am not sure which is better and if it should be different on
basket summary and order receive or the same on both.

QA Comment:
Made a small change to template language; blocking deletion because of items referred to
holds instead.  Changed text to refer to items.
Signed-off-by: Ian Walls <ian.walls@bywatersolutions.com>

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>