Bug 7049: [SIGNED-OFF] Item form on order receive has lots of blank space
The item form on the order receive page (AcqCreateItem = on receive) contains a
lot of blank space which makes it hard to read and fill out.

To test:
- set AcqCreateItem = on receive
- create basket, order something, receive shipment
- check item form is nicely formatted and doesn't include lots of blank space

Note: It will be easier to test if you have an ACQ framework created to hide some
unnecessary subfields, because the hidden fields create the blank space.

Signed-off-by: Magnus Enger <magnus@enger.priv.no>
Created an ACQ framework and hid some of the fields (hidden = 5). Before the
patch there were gaps between the fields as shown in the screenshot from
Katrin. After the patch fields line up nicely, with no extra space between
them.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>