Bug 4161: Preselect active currency for new vendors
Currently when creating a new vendor in acquisitions the alphabetical first
currency will be preselected. For better usability and less mistakes the
active curency should be preselected.

To test:
Check your currency settings in administration.
Your active currency should not be the first in the list.

1) Create a new vendor
Verify the active currency is selected for list and invoice prices.

2) Check vendor summary page
Verify the correct currencies from the vendor record are shown here.

3) Edit vendor
Verify the correct currencies from the vendor record  are shown here.

Signed-off-by: Liz Rea <lrea@nekls.org>
Verified with tests 1, 2, and 3.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>