Bug 7084 - Re-arranging results_summary for tags so that it applies where it is supposed to
To test:
Verify that New Tag text is the same size as all of the other surrounding text.
Verify that it is consistent in size with results that have existing tags.

See screenshot for what it looked like before:
http://screencast.com/t/hnt15kpS7SH9
Notice that the Add button and New tag text were slightly larger than the
similar constructions for results with existing tags. This patch should fix
that.

Signed-off-by: Magnus Enger <magnus@enger.priv.no>
I made sure I had a results list with on record with tags and one without.
Before applying the patch the "New tag" text and "Add" button were
inconsistent for the two records, after applying it they look the same. Nice!

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>