Fix for Bug 7121 - remove comma on items info
Because I can't just take out one comma, I reworked
the way moredetail.pl fetches and outputs biblio
data based on the way it does in detail.pl. This
allows for simpler output in the template.  In the
process I've added subtitle information to the
template.

Revision adds line break for readability

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>