Bug 5533 followup: reinstate Lost status editing
The original patch for 5533 removed the Lost status from the edit items screen; new items were not affected,
and batch changes could still be made, so this was not a consistent change.

This patch removes the filter that hid the Lost status dropdown, but then changes the default MARC framework
to hide this field by default (value -5).  New installations will not see this status in the editor, unless
they manually activate it.  This will enforce the semantics of using the "mark lost" options by default, but
provide a workaround for libraries that understand and accept the consequences of manually editing this field.

To test:

1.  on an existing install, edit an item.  you should have the Lost status visible
2.  on a new install, editing or creating an item should not show the Lost status at all

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>