Bug 6842 Alternate fix, Branch transfer limits broken
The patch which moved the table pager markup into an
include undid the fix which was previously pushed for
this problem. Re-implementing the fix in the pager
include is a simpler fix which doesn't depend on
the Datatables plugin.

The markup change in table-pager.inc fixes the problem
in branch transfer limits and does not affect functionality
on other pages.

The changes in this patch to branch_transfer_limits.tt and
staff-global.css are for cosmetic reasons only.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Saving changes works now correctly. I tried:
- uncheck all for a branch
- check all for a branch
- making changes for different itemtypes on different tabs