follow-up: fixing mistake in database revision number
After long discussion with chris_c (previous RM) and chris_n (previous and actual RMaint), it appears I should not have used 3.06
chris_n will duplicate this entry if it's needed to port this patch onto master.

I renumber the version to 3.07

I also have found in my testings that there is a problem if this DBrev is run twice (once in 3.6 and once in 3.8)
All patrons that are temporarily suspended (ie: have a date in debarred) will become suspended forever
The test was just made on debarred<>0, it's OK if we come from a 3.4, as debarred is a tinyint used as a boolean.
But if we already have a date, it won't do what we want.
I switched to a =1 instead of <>0, that fixes the problem