Bug 7318: Fixes category display in patroncards Patron Search results.
Category description and type were not being sent to the template.  This patch fixes that issue, which also fixes the display.  The display was showing 'Category Description (category type)'.  It now displays 'Category Description (category code) to be consistent with the search label.

This patch also assigns an empty string to $member to get rid of some 'uninitialized' errors in the logs.

Signed-off-by: Liz Rea <wizzyrea@gmail.com>
Confirm this patch fixes the display issue.
Passes prove xt

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>