bug_7458: A call number plugin
Signed-off-by: Melia Meggs <melia@test.bywatersolutions.com>
A library using an inhouse callnumbering system wants to be able to generate the next available call number not already in use.  This custom call number plugin returns the next callnumber not already in use once a correct prefix has been entered.  The prefixes are four to six letters, space, then one or two numbers.