Bug 7633 - Move display of patron image in self checkout
This patch moves the display of the patron image in self-checkout
from above the checkout form into a column to the right of the
form.

The patch also corrects an error in the display of the page title
which was preventing correct display of the library name. To test
this change confirm that the page title shows the value of your
LibraryName system preference or "Koha" if the preference is empty.

http://bugs.koha-community.org/show_bug.cgi?id=7631
Signed-off-by: Liz Rea <wizzyrea@gmail.com>
Confirmed working as expected. Passes template tests.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>