Bug 7408 - Don't cancel found reserves.
At the moment, a reserve will be canceled if it has passed its
expiration date, even if the item is waiting or in the process
of being transferred. The situation could arise where someone
has a hold filled, but it is canceled while in transit, or before
the borrower can pick it up.