Bug 6875 Followup: De-nesting C4::Items breaks bulkmarcimport
Removing 'use C4::Branch' in favour of a subroutine-specific 'require C4::Branch'
causes the GetBranchName subroutine to return an error, as it's not defined
in C4::Items.  Adding "C4::Branch::" scoping fixes the error, which is what's done here.

To confirm problem:
1.  Attempt to run bulkmarcimport.pl before applying the patch.  You should get
    ERROR: Adding items to bib 435 failed: Undefined subroutine &C4::Items::GetBranchName
    called at /home/sekjal/kohaclone/C4/Items.pm line 656, <GEN13> line 435.

To test:
2.  apply patch
3.  run bulkmarcimport again.  Error should disappear.

Signed-off-by: Ian Walls <koha.sekjal@gmail.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>
Note that this problem only appears when importing records with item (952)
fields.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>