Bug 7760: Add unique ids and a module class to all administration pages
This is the first patch for bug 7760 and touches all pages in administration.
This adds a unique id "admin_<filename>" and a class "admin" to the body tag of
each page in administration.

Note: aqcontract can only be accessed from the acquisition module, so I made it acq
instead of admin.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>