Bug 7727 - NORMARC XSLT OPAC detail view shows double tabs
With marcflavour = NORMARC and OPACXSLTDetailsDisplay = Using XSLT,
the detail view displays a double set of tabs for Normal/MARC/ISBD
views, one above the title and one below.

To test:
marcflavour = NORMARC
OPACXSLTDetailsDisplay = Using XSLT

Since the contents of the recaords do not matter here, the patch
can be tested on MARC records of any flavour.

- Check that double tabs are shown
- Apply the patch
- Check that only the set of tabs at the top is shown

Signed-off-by: Marijana Glavica <mglavica@ffzg.hr>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>