Bug 7821 - {langcode} will be replaced with current interface language
This makes bug 4032 support multi-language Koha installation which is important
for some of Koha users.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
1) After update XSLT system preferences are set to 'default'. All views
(OPAC and staff, results and detail page) work and switching languages
selects the correct translated file.

2) Leaving empty works and normal views are used.

3) Putting in a single URL without placeholder works.

4) Putting in a URL with placeholder works. Created local directories
have to match the language codes.

NOTE: In order to make the URLs and patchs work, the Utils file needs to be
copied too.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>